Create everything but sliders
Initial build-out of the code framework, which has been tested.  Have not integrated the NeoSliders, but it works with the current constants and a switch.  Would also love to build in some better logic around the randomizer to make it "more random" (read: eliminate the chance of getting the same animation twice), but baby steps.